requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
gunicorn==21.2.0
requests==2.31.0
geopy==2.4.1
folium==0.15.1
pandas==2.2.2